+++ src/App.vue
@@ -1,5 +1,5 @@
 <script lang="ts" setup>
-import Snackbar from '@/components/Snackbar.vue';
+import Snackbar from '@/components/etc/Snackbar.vue';
 </script>
 
 <template>
