+++ src/components/chat/config/ChatTypeSelector.vue
