+++ src/components/chat/config/OpenaiApiKeyGuide.vue
