+++ src/components/chat/config/OpenaiModelSelector.vue
